convert alias file from csv to json
